refactor: move UI helpers into controls package
