+++ README.md
@@ -1,4 +1,4 @@
-version: 0.2.12
+version: 0.2.13
 Author: Miguel Marina <karel.capek.robotics@gmail.com> - [LinkedIn](https://www.linkedin.com/in/progman32/)
 # AutoML
 
@@ -1642,6 +1642,7 @@ and run the build again if you hit this issue.
 
 
 ## Version History
+- 0.2.13 - Consolidate button helpers and message dialogs under controls package.
 - 0.2.12 - Fix property initialisation error in safety analysis facade.
 - 0.2.11 - Keep splash screen visible through startup and five-second post-load delay.
 - 0.2.10 - Hold splash screen for five seconds after initialisation completes.
