+++ gui/dialogs/user_select_dialog.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import tkinter as tk
 from tkinter import ttk, simpledialog
 
-from gui.mac_button_style import apply_purplish_button_style
+from gui.controls.mac_button_style import apply_purplish_button_style
 
 
 class UserSelectDialog(simpledialog.Dialog):
