+++ gui/gsn_explorer.py
@@ -5,11 +5,10 @@ import tkinter as tk
 from tkinter import ttk, simpledialog
 
 from mainappsrc.models.gsn import GSNNode, GSNDiagram, GSNModule
-from gui import format_name_with_phase, messagebox
+from gui import format_name_with_phase
+from gui.controls import messagebox
 from gui.style_manager import StyleManager
 from gui.icon_factory import create_icon
-from .name_utils import collect_work_product_names, unique_name_v4
-from . import messagebox
 
 
 class GSNExplorer(tk.Frame):
