+++ gui/safety_management_explorer.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 
 import tkinter as tk
 from tkinter import ttk, simpledialog
-from gui import messagebox, format_name_with_phase
+from gui.controls import messagebox
+from gui import format_name_with_phase
 from mainappsrc.models.sysml.sysml_repository import SysMLRepository
 from dataclasses import dataclass, field
 from typing import List, Dict
