+++ gui/safety_management_toolbox.py
@@ -14,8 +14,9 @@ from analysis.models import (
     global_requirements,
 )
 from gui.architecture import GovernanceDiagramWindow
-from gui import messagebox, add_treeview_scrollbars
-from gui.mac_button_style import apply_translucid_button_style
+from gui.controls import messagebox
+from gui import add_treeview_scrollbars
+from gui.controls.mac_button_style import apply_translucid_button_style
 from gui.icon_factory import create_icon
 from mainappsrc.models.sysml.sysml_repository import SysMLRepository
 from gui.toolboxes import configure_table_style, _wrap_val, enable_treeview_reorder
