+++ gui/search_toolbox.py
@@ -5,8 +5,8 @@ import re
 import tkinter as tk
 from tkinter import ttk
 
-from gui import messagebox
-from gui.mac_button_style import apply_translucid_button_style
+from gui.controls import messagebox
+from gui.controls.mac_button_style import apply_translucid_button_style
 
 # Additional model sections that can be searched.  Each tuple contains a
 # human-readable category name, the name of a method on the ``app`` object
