+++ gui/style_editor.py
@@ -1,6 +1,6 @@
 import tkinter as tk
 from tkinter import colorchooser, filedialog
-from gui import messagebox
+from gui.controls import messagebox
 from .style_manager import StyleManager
 
 class StyleEditor(tk.Toplevel):
