+++ mainappsrc/fta_subapp.py
@@ -6,7 +6,7 @@ import re
 
 from analysis.fmeda_utils import GATE_NODE_TYPES
 from config.automl_constants import dynamic_recommendations, VALID_SUBTYPES
-from gui import messagebox
+from gui.controls import messagebox
 try:  # pragma: no cover - support direct module import
     from .models.fta.fault_tree_node import FaultTreeNode
 except Exception:  # pragma: no cover
