+++ tests/test_diagram_clipboard_no_focus.py
@@ -7,7 +7,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
 from AutoML import AutoMLApp
 from gui.architecture import SysMLDiagramWindow, _get_next_id, ARCH_WINDOWS, SysMLObject
-from gui import messagebox
+from gui.controls import messagebox
 
 
 class DummyRepo:
