+++ tests/test_gsn_explorer.py
@@ -4,7 +4,7 @@ import types
 
 from tkinter import simpledialog
 from mainappsrc.models.gsn import GSNNode, GSNDiagram, GSNModule
-from gui import messagebox
+from gui.controls import messagebox
 from gui.gsn_explorer import GSNExplorer
