+++ tests/test_listbox_hover_highlight.py
@@ -4,7 +4,7 @@ import tkinter as tk
 import pytest
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
-from gui.button_utils import enable_listbox_hover_highlight
+from gui.controls.button_utils import enable_listbox_hover_highlight
 
 
 def _rgb(value):
