+++ tests/test_listbox_hover_invalid_widget.py
@@ -2,7 +2,7 @@ import sys
 from pathlib import Path
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
-from gui.button_utils import enable_listbox_hover_highlight
+from gui.controls.button_utils import enable_listbox_hover_highlight
 
 
 class DummyRoot:
