+++ tests/test_treeview_hover_highlight.py
@@ -5,7 +5,7 @@ from tkinter import ttk
 import pytest
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
-from gui.button_utils import enable_listbox_hover_highlight
+from gui.controls.button_utils import enable_listbox_hover_highlight
 
 
 def _rgb(value):
